requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
pandas>=2.2.0
gspread>=6.0.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.97.0
python-dotenv>=1.0.0
schedule>=1.2.0
openpyxl>=3.1.0
numpy>=1.26.0 